fix: production audit — 100 fixes across correctness, security, performance, and testing
P0 Critical:
- Define TEMPLATE_SCHEMA_VERSION preventing ImportError on template operations
- Return sweep mode and request_snapshot in API responses (genetic sweeps showed as grid)
- Add Decimal fallbacks for NaN on NOT NULL summary columns
- Fix exception re-raising in market data coalesced fetch
- Remove billing gates from analysis read/delete endpoints

P1 High:
- Filter bars with close_price <= 0 preventing division-by-zero in BSM
- Bounded S3 read preventing OOM on corrupted ContentLength
- Release pipeline lock before retry (nightly retries were ineffective)
- Remove skipped from CD e2e allowed-conclusions
- Use time.monotonic() in rate limiter
- Fix Sortino ratio denominator from (n-1) to n
- Split migration constraint VALIDATE into post-deploy step
- Fix backfill_metrics infinite reprocessing of all-win runs
- Fix Content-Disposition header injection via unescaped quotes
- Hash-based sweep duplicate detection

P2 Medium:
- DLQ Redis invalidation, DataUnavailableError 422->503, S3 jitter
- SSE Last-Event-ID support, heartbeat timeout 90s->45s
- Recursive billing event redaction, updated_at on error recovery
- Sweep DTE integer validation, equity_curve null safety, quota refresh
- Circuit breaker atomic INCR+EXPIRE, pool gauge logging
- Concurrent scan limit, evaluated_candidate_count fix
- Audit events for backtest/sweep/scan/export/template CRUD
- GDPR export total counts, IntegrityError differentiation

Tests:
- Rewrite 6 empty/fake tests with real production code coverage
- Fix auth bypass, injection, lifecycle, ranking tests
- Add losing trade scenarios to 6 strategy test classes

P3:
- BSM dividend yield, regime zero return fix, Sharpe/Sortino min sample
- Score formula fix, wheel margin warning, jade lizard max_loss
- Custom legs mixed sides, win rate normalization, 180-day quick backtest
- Configurable billing grace periods, reconciliation automation
- Pool exhaustion alerting, hash_ip caching, validation bounds alignment
- JWT leeway config, RSA key validation, client pagination
- NaN/Inf quote guard, volume/shares_per_contract int types

Made-with: Cursor

+++ src/backtestforecast/market_data/types.py
@@ -12,7 +12,7 @@ class DailyBar:
     high_price: float
     low_price: float
     close_price: float
-    volume: float
+    volume: int
 
 
 @dataclass(frozen=True, slots=True)
@@ -20,8 +20,8 @@ class OptionContractRecord:
     ticker: str
     contract_type: str
     expiration_date: date
-    strike_price: float
-    shares_per_contract: float
+    strike_price: float  # Precision note: float cannot represent all decimals exactly. Use Decimal(str(strike_price)) at financial computation boundaries.
+    shares_per_contract: int
 
 
 @dataclass(frozen=True, slots=True)
